requirements.txt
PyYAML==3.10
path.py==3.0.1
Jinja2==2.6
Flask==0.9
Werkzeug==0.8.3
Flask-Script==0.5.3
